pyproject.toml
[tool.poetry]
name = "registrationbot-school"
version = "0.1.0"
description = ""
authors = ["amadeo <risefenixpuryfire@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telebot = "^0.0.5"
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.7.1"
pyyaml = "^6.0.2"
aiohttp = "^3.11.12"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
